Added service layer
