+++ src/main/java/com/tavisca/workshops/todoapp/controllers/TodoController.java
@@ -1,66 +1,42 @@
 package com.tavisca.workshops.todoapp.controllers;
 
-import org.json.JSONObject;
-import org.springframework.http.HttpStatus;
+import com.tavisca.workshops.todoapp.services.TodoService;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.time.Instant;
 import java.util.ArrayList;
 import java.util.List;
 
 @RestController
 @CrossOrigin(origins = "*", allowedHeaders = "*")
 public class TodoController {
+    @Autowired
+    TodoService todoService;
     private List<String> todos = new ArrayList<>();
 
     @GetMapping(path = "/todos")
-    public ResponseEntity<?> getTodos() {
-        return sendResponse("All todos retrieved");
+    public ResponseEntity<?> get() {
+        return todoService.getTodos();
     }
 
     @GetMapping(path = "/todos/{id}")
-    public ResponseEntity<?> getTodo(@PathVariable("id") int todoid) {
-        if (todoid >= todos.size())
-            return new ResponseEntity<>("", HttpStatus.BAD_REQUEST);
-        JSONObject jsonResponse = new JSONObject()
-                .put("todoname", todos.get(todoid))
-                .put("status", "Todo " + (todoid + 1) + " retrieved")
-                .put("timestamp", Instant.now().toString());
-        return new ResponseEntity<>(jsonResponse.toString(), HttpStatus.OK);
+    public ResponseEntity<?> getById(@PathVariable("id") int todoid) {
+        return todoService.getTodoById(todoid);
     }
 
     @PostMapping(path = "/todos")
-    public ResponseEntity<?> addTodo(@RequestBody String json) {
-        JSONObject obj = new JSONObject(json);
-        String todoname = obj.getString("todoname");
-        todos.add(todoname);
-        return sendResponse("Todo " + todoname + " Added");
+    public ResponseEntity<?> add(@RequestBody String json) {
+        return todoService.addTodo(json);
     }
 
     @PutMapping(path = "/todos/{id}")
-    public ResponseEntity<?> updateTodo(@PathVariable("id") int todoid, @RequestBody String json) {
-        JSONObject jsonObject = new JSONObject(json);
-        if (todoid >= todos.size())
-            return new ResponseEntity<>("", HttpStatus.BAD_REQUEST);
-        String todoname = jsonObject.getString("todoname");
-        todos.set(todoid, todoname);
-        return sendResponse("Todo-" + (todoid + 1) + " Updated");
+    public ResponseEntity<?> update(@PathVariable("id") int todoid, @RequestBody String json) {
+        return todoService.updateTodo(todoid, json);
     }
 
     @DeleteMapping(path = "/todos/{id}")
     public ResponseEntity<?> deleteTodo(@PathVariable("id") int todoid) {
-        if (todoid >= todos.size())
-            return new ResponseEntity<>("", HttpStatus.BAD_REQUEST);
-        todos.remove(todoid);
-        return sendResponse("Todo-" + (todoid + 1) + " Deleted");
-    }
-
-    private ResponseEntity<?> sendResponse(String status) {
-        JSONObject jsonResponse = new JSONObject()
-                .put("todos", todos)
-                .put("status", status)
-                .put("timestamp", Instant.now().toString());
-        return new ResponseEntity<>(jsonResponse.toString(), HttpStatus.OK);
+        return todoService.deleteTodo(todoid);
     }
 }
